requirements.txt
click>=8.0.0
pyyaml>=6.0.0
colorama>=0.4.6
python-dateutil>=2.8.2 